Update stats.clj "rolling-window-set" function, exchange the real argument "num-buckets" and "s" of "rolling-window" function
(defn rolling-window-set [updater merger extractor num-buckets & bucket-sizes]
(RollingWindowSet. updater extractor (dofor [s bucket-sizes] (rolling-window updater merger extractor s num-buckets)) nil)
)
(defrecord RollingWindow [updater merger extractor bucket-size-secs num-buckets buckets])
if not exchange the real argument ”num-buckets“ and "s" of “rolling-window” function, then the "bucket-size-secs" of RollingWindow is 30/540/4320, and the "num-buckets" of RollingWindow is 20
I think that the "bucket-size-secs" of RollingWindow is 20, and the "num-buckets" of RollingWindow is 30/540/4320.