configurable supervisor log filename
For the storm-on-mesos project there are multiple Storm Supervisors on
every worker host.  Each of these Supervisors is dedicated to a
particular topology.  A challenge with this design is that the shared
supervisor log file contains logs from every single topology that
has run on a particular host.  We instead want the option to have
separate supervisor log files for each topology, so that topology owners
can more easily debug their own problems.

This change allows having configurable log files for every launch of
the Supervisor, since we will be able to embed the topology ID into the
supervisor log filename.

Notably, storm-0.10 already includes a similar change for the worker
log filenames (came in with Yahoo!'s multi-tenant changes).

Conflicts:
	bin/storm.py