off-by-one error for UNCOMMITTED_LATEST and UNCOMMITTED_EARLIEST on initialize
In the source code documentation, it says for both UNCOMMITTED_LATEST
and UNCOMMITTED_EARLIEST it will fetch the last committed offset ( if it
exists ).  But the code fetches last committed + 1;

If a topology is killed/crashes and then is redeployed, this off-by-one
error causes one message per partition to be dropped.