Fix emitting of tuples from bolts with Python 3
In Python 3, the map() function no longer returns a list, but instead an iterable object, which is not JSON serializable.  To fix this, convert the result of the map() call to a list with the list() function.  This is still compatible with Python 2 (although will create a copy of the list).