STORM-2614: Enhance stateful windowing to persist the window state
Right now the tuples in window are stored in memory. This limits the usage to windows
that fit in memory. Also the source tuples cannot be acked until the window expiry.
By persisting the window transparently in the state backend and caching/iterating them as needed,
we could support larger windows and also support windowed bolts with user/application state.