[STORM-2892] fix PATH substitution in flux tests
The tests fail when the PATH environment variable has a trailing colon, despite that being a valid PATH.
This happens because it is substituted directly into the resultant YAML file, which results in invalid
YAML (since you cannot end a map's value with ":" in the raw text).

So the solution is to wrap the map value with double-quotes.

Also fix 2 typos in comments in this file.