Setup ShellBolt logger in the thread using it
Prior to this the logger set up would occur in a separate thread (the
parent), to the one doing the logging. This meant that MDC values, which
are thread local, could not persist from the set up phase. By doing set
up in the same thread, MDC values are now reliably available during log
calls.