[STORM-3401] Add OpenLab CI configuration for ARM64 build
This patch adds the CI configuration to enable the support for arm
build in OpenLab.

After this, each pull request in theopenlab/storm will trigger the
storm-arm64-build job which verified the arm build on ARM cluster.

As the first version, we mark this job as non-voting job, that means
the build job result is just a reference for developer and would not
block the PR to be merged. Once we make sure the job can be executed
stable, then we can mark it as normal "voting" job.